improve React reconciliation performance with new view components
