+++ frontend/src/components/Results-View/Item-Menu/Person-View/Person-Card/Item-View/Item-Row/index.tsx
