+++ frontend/src/components/Results-View/Item-Menu/Person-View/Person-Card/Person-Actions/person-actions.component.css
