+++ frontend/src/components/Results-View/Item-Menu/Person-View/Person-Card/index.tsx
