+++ frontend/src/components/Results-View/Item-Menu/Person-View/Person-Card/person-card.component.css
