+++ frontend/src/components/Results-View/Item-Menu/Person-View/person-view.component.tsx
@@ -1,0 +1,15 @@
+import { observer } from "mobx-react";
+import { PersonMobx } from "../../../../models/person.model";
+import { PersonCard } from "./Person-Card";
+
+const PersonsView = observer(({ persons }: { persons: PersonMobx[] }) => {
+  return (
+    <>
+      {persons.map((person: PersonMobx) => (
+        <PersonCard key={person.id} person={person} />
+      ))}
+    </>
+  );
+});
+
+export { PersonsView };
